+++ src/Button.js
@@ -110,7 +110,7 @@ Button.propTypes = {
    * If enabled, any children button will be rendered as actions, remember to provide an icon.
    * @default vertical. This will disable any onClick function from being called on the main button.
    */
-  fab: PropTypes.oneOf(['vertical', 'horizontal']),
+  fab: PropTypes.oneOf(['vertical', 'horizontal', 'toolbar']),
   /**
    * The icon to display, if specified it will create a button with the material icon.
    */